.gitignore: Removed restriction on access to Data folder Added restriction on access to R working folder
Deleted version of generate_GEM_GDX.gms from root folder

Deleted i_NrgDemand_df.csv from root folder. This was an output from an earlier run that shouldn't have gone in this folder.

Added annual energy forecast dataset (2012-50)

Added i_NrgDemand.gdx file. This is the final output of the generate_gem_GDX code. It is one of the inputs into the GEM model.

Minor changes to convert_CSV_to_GDx.R code. Needed to explicitly move the working directory to that of the GMS file before running GAMS (and then returning back to the working directory of the project).

Minor changes to generate_gem_GDX.R. The code now looks in the config file for the path to the GAMS API. Commented out some sections.