Added new code that generates a single GDX input file from the individual input CSVs (GDXfromCSVs.R). Currently allows for a different demand file to be used. More functionality to come in the future.
Changed the code that used to create a single demand GDX to just creating a demand which CSV which can be used later as input to the overall GDX file (createFinalCSV.R).

Added default CSVs for all inputs (set, subsets and parameters) to the Data folder.

Renamed/moved some files